+++ Persistence/src/main/java/com/kalaazu/persistence/repository/SkilltreeUnlocksRepository.java
@@ -13,5 +13,5 @@ import org.springframework.stereotype.Repository;
  * @author Manulaiko <manulaiko@gmail.com>
  */
 @Repository
-public interface SkilltreeUnlocksRepository extends JpaRepository<SkilltreeUnlocksEntity, Integer> {
+public interface SkilltreeUnlocksRepository extends JpaRepository<SkilltreeUnlocksEntity, Short> {
 }
