+++ Persistence/src/main/java/com/kalaazu/persistence/service/MapsPortalsServiceImpl.java
@@ -32,7 +32,7 @@ public class MapsPortalsServiceImpl implements MapsPortalsService {
      * @inheritDoc
      */
     @Override
-    public MapsPortalsEntity find(Integer id) {
+    public MapsPortalsEntity find(Byte id) {
         return this.repository.findById(id).orElse(null);
     }
 
@@ -56,7 +56,7 @@ public class MapsPortalsServiceImpl implements MapsPortalsService {
      * @inheritDoc
      */
     @Override
-    public boolean delete(Integer id) {
+    public boolean delete(Byte id) {
         this.repository.deleteById(id);
 
         return !this.repository.existsById(id);
